LayoutTests/platform/ios-sim-deprecated/compositing/columns/hittest-composited-in-paginated-expected.txt

CONSOLE MESSAGE: line 12: TypeError: null is not an object (evaluating 'document.elementFromPoint(500, 200).classList')
Layer tree goes here in DRT
